server/requirements.txt
mongoengine==0.13.0
pymongo==3.5.1
python-dateutil-2.6.1
requests==2.9.1
trueskill=0.4.4